+++ packages/catalog-model/src/kinds/ComponentEntityV1alpha1.ts
@@ -29,8 +29,10 @@ const schema = yup.object<Partial<ComponentEntityV1alpha1>>({
       type: yup.string().required().min(1),
       lifecycle: yup.string().required().min(1),
       owner: yup.string().required().min(1),
+      subcomponentOf: yup.string().notRequired().min(1),
       providesApis: yup.array(yup.string().required()).notRequired(),
       consumesApis: yup.array(yup.string().required()).notRequired(),
+      system: yup.string().notRequired().min(1),
     })
     .required(),
 });
@@ -42,8 +44,10 @@ export interface ComponentEntityV1alpha1 extends Entity {
     type: string;
     lifecycle: string;
     owner: string;
+    subcomponentOf?: string;
     providesApis?: string[];
     consumesApis?: string[];
+    system?: string;
   };
 }