+++ plugins/catalog-import/src/util/useGithubRepos.ts
@@ -18,22 +18,25 @@ import * as YAML from 'yaml';
 import { useApi, configApiRef } from '@backstage/core';
 import { catalogImportApiRef } from '../api/CatalogImportApi';
 import { ConfigSpec } from '../components/ImportComponentPage';
-import parseGitUri from 'git-url-parse';
+import parseGitUrl from 'git-url-parse';
 
 // TODO: (O5ten) Refactor into a core API instead of direct usage like this
 // https://github.com/backstage/backstage/pull/3613#issuecomment-7408929430
-import { readGitHubIntegrationConfigs } from '@backstage/integration';
+import {
+  GitHubIntegrationConfig,
+  readGitHubIntegrationConfigs,
+} from '@backstage/integration';
 
 export function useGithubRepos() {
   const api = useApi(catalogImportApiRef);
   const config = useApi(configApiRef);
 
-  const submitPrToRepo = async (selectedRepo: ConfigSpec) => {
+  const getGithubIntegrationConfig = (location: string) => {
     const {
       name: repoName,
       owner: ownerName,
       resource: hostname,
-    } = parseGitUri(selectedRepo.location);
+    } = parseGitUrl(location);
 
     const configs = readGitHubIntegrationConfigs(
       config.getOptionalConfigArray('integrations.github') ?? [],
@@ -41,9 +44,22 @@ export function useGithubRepos() {
     const githubIntegrationConfig = configs.find(v => v.host === hostname);
     if (!githubIntegrationConfig) {
       throw new Error(
-        `Unable to locate github-integration for repo-location: ${selectedRepo.location}`,
+        `Unable to locate github-integration for repo-location, ${location}`,
       );
     }
+    return {
+      repoName,
+      ownerName,
+      githubIntegrationConfig,
+    };
+  };
+
+  const submitPrToRepo = async (selectedRepo: ConfigSpec) => {
+    const {
+      repoName,
+      ownerName,
+      githubIntegrationConfig,
+    } = getGithubIntegrationConfig(selectedRepo.location);
     const submitPRResponse = await api
       .submitPrToRepo({
         owner: ownerName,
@@ -54,7 +70,7 @@ export function useGithubRepos() {
         githubIntegrationConfig,
       })
       .catch(e => {
-        throw new Error(`Failed to submit PR to repo:\n${e.message}`);
+        throw new Error(`Failed to submit PR to repo, ${e.message}`);
       });
 
     await api
@@ -62,14 +78,41 @@ export function useGithubRepos() {
         location: submitPRResponse.location,
       })
       .catch(e => {
-        throw new Error(`Failed to create repository location:\n${e.message}`);
+        throw new Error(`Failed to create repository location, ${e.message}`);
       });
 
     return submitPRResponse;
   };
 
+  const checkForExistingCatalogInfo = async (
+    location: string,
+  ): Promise<{ exists: boolean; url?: string }> => {
+    let githubConfig: {
+      repoName: string;
+      ownerName: string;
+      githubIntegrationConfig: GitHubIntegrationConfig;
+    };
+    try {
+      githubConfig = getGithubIntegrationConfig(location);
+    } catch (e) {
+      return { exists: false };
+    }
+    return await api
+      .checkForExistingCatalogInfo({
+        owner: githubConfig.ownerName,
+        repo: githubConfig.repoName,
+        githubIntegrationConfig: githubConfig.githubIntegrationConfig,
+      })
+      .catch(e => {
+        throw new Error(
+          `Failed to inspect repository for existing catalog-info.yaml, ${e.message}`,
+        );
+      });
+  };
+
   return {
     submitPrToRepo,
+    checkForExistingCatalogInfo,
     generateEntityDefinitions: (repo: string) =>
       api.generateEntityDefinitions({ repo }),
     addLocation: (location: string) =>