+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/github.test.ts
@@ -17,15 +17,10 @@
 jest.mock('@octokit/rest');
 jest.mock('./helpers');
 
-import { Octokit } from '@octokit/rest';
-import {
-  OctokitResponse,
-  ReposCreateInOrgResponseData,
-  UsersGetByUsernameResponseData,
-} from '@octokit/types';
+import { getVoidLogger } from '@backstage/backend-common';
+import { Octokit, RestEndpointMethodTypes } from '@octokit/rest';
 import { GithubPublisher } from './github';
 import { initRepoAndPush } from './helpers';
-import { getVoidLogger } from '@backstage/backend-common';
 
 const { mockGithubClient } = require('@octokit/rest') as {
   mockGithubClient: {
@@ -54,12 +49,12 @@ describe('GitHub Publisher', () => {
           data: {
             clone_url: 'https://github.com/backstage/backstage.git',
           },
-        } as OctokitResponse<ReposCreateInOrgResponseData>);
+        } as RestEndpointMethodTypes['repos']['createInOrg']['response']);
         mockGithubClient.users.getByUsername.mockResolvedValue({
           data: {
             type: 'Organization',
           },
-        } as OctokitResponse<UsersGetByUsernameResponseData>);
+        } as RestEndpointMethodTypes['users']['getByUsername']['response']);
 
         const result = await publisher.publish({
           values: {
@@ -104,12 +99,12 @@ describe('GitHub Publisher', () => {
           data: {
             clone_url: 'https://github.com/backstage/backstage.git',
           },
-        } as OctokitResponse<ReposCreateInOrgResponseData>);
+        } as RestEndpointMethodTypes['repos']['createForAuthenticatedUser']['response']);
         mockGithubClient.users.getByUsername.mockResolvedValue({
           data: {
             type: 'User',
           },
-        } as OctokitResponse<UsersGetByUsernameResponseData>);
+        } as RestEndpointMethodTypes['users']['getByUsername']['response']);
 
         const result = await publisher.publish({
           values: {
@@ -148,12 +143,12 @@ describe('GitHub Publisher', () => {
         data: {
           clone_url: 'https://github.com/backstage/backstage.git',
         },
-      } as OctokitResponse<ReposCreateInOrgResponseData>);
+      } as RestEndpointMethodTypes['repos']['createForAuthenticatedUser']['response']);
       mockGithubClient.users.getByUsername.mockResolvedValue({
         data: {
           type: 'User',
         },
-      } as OctokitResponse<UsersGetByUsernameResponseData>);
+      } as RestEndpointMethodTypes['users']['getByUsername']['response']);
 
       const result = await publisher.publish({
         values: {
@@ -205,12 +200,12 @@ describe('GitHub Publisher', () => {
         data: {
           clone_url: 'https://github.com/backstage/backstage.git',
         },
-      } as OctokitResponse<ReposCreateInOrgResponseData>);
+      } as RestEndpointMethodTypes['repos']['createInOrg']['response']);
       mockGithubClient.users.getByUsername.mockResolvedValue({
         data: {
           type: 'Organization',
         },
-      } as OctokitResponse<UsersGetByUsernameResponseData>);
+      } as RestEndpointMethodTypes['users']['getByUsername']['response']);
 
       const result = await publisher.publish({
         values: {
@@ -254,12 +249,12 @@ describe('GitHub Publisher', () => {
         data: {
           clone_url: 'https://github.com/backstage/backstage.git',
         },
-      } as OctokitResponse<ReposCreateInOrgResponseData>);
+      } as RestEndpointMethodTypes['repos']['createForAuthenticatedUser']['response']);
       mockGithubClient.users.getByUsername.mockResolvedValue({
         data: {
           type: 'User',
         },
-      } as OctokitResponse<UsersGetByUsernameResponseData>);
+      } as RestEndpointMethodTypes['users']['getByUsername']['response']);
 
       const result = await publisher.publish({
         values: {