+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/cli
 
+## 0.4.6
+
+### Patch Changes
+
+- 94fdf4955: Get rid of all usages of @octokit/types, and bump the rest of the octokit dependencies to the latest version
+- 08e9893d2: Handle no npm info
+- 9cf71f8bf: Added experimental `create-github-app` command.
+
 ## 0.4.5
 
 ### Patch Changes
@@ -21,7 +29,7 @@
 
 ### Patch Changes
 
-- 19554f6d6: Added Github Actions for Create React App, and allow better imports of files inside a module when they're exposed using `files` in `package.json`
+- 19554f6d6: Added GitHub Actions for Create React App, and allow better imports of files inside a module when they're exposed using `files` in `package.json`
 - 7d72f9b09: Fix for `app.listen.host` configuration not properly overriding listening host.
 
 ## 0.4.2