+++ plugins/scaffolder/CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### Patch Changes
 
-- 19554f6d6: Added Github Actions for Create React App, and allow better imports of files inside a module when they're exposed using `files` in `package.json`
+- 19554f6d6: Added GitHub Actions for Create React App, and allow better imports of files inside a module when they're exposed using `files` in `package.json`
 - Updated dependencies [1dc445e89]
 - Updated dependencies [342270e4d]
   - @backstage/core@0.4.2